restructure main file inside of the cmd package
